Walk up directory tree to find config files.
Add support for searching for google repo and .gitmulti files by walking
up the directory tree. This allows for running git-multi from any
directory below the topdir that contains either the google repo manifest
or the .gitmulti config files.

This does not support running git-multi from within a hierarchy that
relies on looking for for a bunch of parallel git repos with all .git
dirs at the same depth. For that, you still to cd to the parent of the
git repo you are currently in.

Signed-off-by: Theodore A. Roth <troth@openavr.org>